Fixes and Enhancements to StatesPlugin
A bunch of changes.  Refactored file name to something more descriptive.  Refactored some of the requestHandler code into submethods for reuse and to make it more bite sized.  Fixed issue with text not rotating to face camera.  Fixed issue with text not being set on program restart.  Fixed issue with text not setting it's height properly relative to the size of the mini it's above.  Fixed issue with text not working with CMP minis properly (not getting the SkinnedMeshRenderer height value properly).

Im sure there are other things also.  Lot's of good stuff (I hope)

-A